Refactor pipeline with extract, model, and transform folders
