+++ uniflow/constants.py
@@ -1,0 +1,4 @@
+"""Flow constants."""
+
+ROOT_NAME = "root"
+OUTPUT_NAME = "output"
