--- uniflow/extract/extract_txt_flow.py
@@ -1,30 +1,0 @@
-"""Extract txt flow."""
-
-from typing import Sequence
-
-from uniflow.flow.flow import Flow
-from uniflow.node.node import Node
-from uniflow.op.extract.txt_op import ExtractTxtOp, ProcessTxtOp
-
-
-class ExtractTxtFlow(Flow):
-    """Extract txt Flow Class."""
-
-    def __init__(self) -> None:
-        """Extract txt Flow Constructor."""
-        super().__init__()
-        self._extract_txt_op = ExtractTxtOp(name="extract_txt_op")
-        self._process_txt_op = ProcessTxtOp(name="process_txt_op")
-
-    def run(self, nodes: Sequence[Node]) -> Sequence[Node]:
-        """Run Extract txt Flow.
-
-        Args:
-            nodes (Sequence[Node]): Nodes to run.
-
-        Returns:
-            Sequence[Node]: Nodes after running.
-        """
-        nodes = self._extract_txt_op(nodes)
-        nodes = self._process_txt_op(nodes)
-        return nodes
