--- uniflow/flow/constants.py
@@ -1,4 +1,0 @@
-"""Flow constants."""
-
-ROOT_NAME = "root"
-OUTPUT_NAME = "output"
