+++ uniflow/model/flow/__init__.py
@@ -1,0 +1,18 @@
+"""Flow __init__ module."""
+# this register all possible flow into FlowFactory through
+# FlowFactory.register(cls.__name__, cls) in Flow __init_subclass__
+from uniflow.flow import LinearFlow  # noqa: F401
+from uniflow.model.flow.model_flow import (  # noqa: F401;
+    BaseModelFlow,
+    HuggingFaceModelFlow,
+    LMQGModelFlow,
+    OpenAIModelFlow,
+)
+
+__all__ = [
+    "BaseModelFlow",
+    "HuggingFaceModelFlow",
+    "LinearFlow",
+    "LMQGModelFlow",
+    "OpenAIModelFlow",
+]
