+++ uniflow/transform/config.py
@@ -1,0 +1,11 @@
+"""Transform config module."""
+
+from dataclasses import dataclass
+
+
+@dataclass
+class Config:
+    """Transform Config Class."""
+
+    flow_name: str = "TransformFlow"
+    num_thread: int = 1
